refactor daemon executor attempt ownership
